requirements.txt
streamlit==1.24.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
requests==2.31.0
Pillow==9.5.0